requirements.txt
mysql-connector-python
Faker
python-dotenv